+++ Tests/Obfy.Tests/SourceObfuscatorTests.cs
@@ -47,6 +47,8 @@ public class SourceObfuscatorTests
         var newSource = context.Compilation!.SyntaxTrees.First().ToString();
         newSource.ShouldNotContain("\"Hello World\"");
         newSource.ShouldContain("__ObfyStringDecryptor");
+        GetErrors(context.Compilation!).ShouldBeEmpty();
+        EmitAndInvoke(context.Compilation!, "Test", "GetMessage").ShouldBe("Hello World");
     }
 
     [Fact]
@@ -72,6 +74,8 @@ public class SourceObfuscatorTests
         var newSource = context.Compilation!.SyntaxTrees.First().ToString();
         newSource.ShouldNotContain("Hello ");
         newSource.ShouldContain("__ObfyStringDecryptor");
+        GetErrors(context.Compilation!).ShouldBeEmpty();
+        EmitAndInvoke(context.Compilation!, "Test", "Greet", "Ada").ShouldBe("Hello Ada world");
     }
 
     [Fact]
@@ -315,6 +319,8 @@ public class SourceObfuscatorTests
 
         var newSource = context.Compilation!.SyntaxTrees.First().ToString();
         newSource.ShouldNotContain(@"@""C:\Users\Test\Documents""");
+        GetErrors(context.Compilation!).ShouldBeEmpty();
+        EmitAndInvoke(context.Compilation!, "Test", "GetPath").ShouldBe(@"C:\Users\Test\Documents");
     }
 
     [Fact]